Moved Axis,Button,Pov to joystick & fixed porblems
